--- a/connectivity/check/check.go
+++ b/connectivity/check/check.go
@@ -51,6 +51,12 @@ type Parameters struct {
 	PerfHostNet            bool
 	PerfPodNet             bool
 	PerfSamples            int
+	PerfMessageSize        int
+	PerfMixed              bool
+	PerfThroughput         bool
+	PerfCRR                bool
+	PerfRR                 bool
+	PerfUDP                bool
 	CurlImage              string
 	PerformanceImage       string
 	JSONMockImage          string