--- a/connectivity/perf/common/metrics.go
+++ b/connectivity/perf/common/metrics.go
@@ -99,6 +99,7 @@ type PerfTests struct {
 	SameNode bool
 	Scenario string
 	Sample   int
+	MsgSize  int
 	Duration time.Duration
 }