This PR contains various improvements to net perf testing - more details in commit messages.

Most notable improvement is that we no longer use fixed message size for TCP_STREAM test, which results in way more reasonable results as compared to the previous tests. 
First of all, we can finally saturate network throughput and also throughput between pods on the same node is way higher than throughput between pods running on different nodes. Example test results for native routing:
```
-------------------------------------------------------------------------------------
📋 Scenario        | Node       | Test            | Duration        | Throughput Mb/s 
-------------------------------------------------------------------------------------
📋 pod-to-pod      | same-node  | TCP_STREAM      | 10s             | 14257.07     
📋 host-to-host    | same-node  | TCP_STREAM      | 10s             | 25766.59     
📋 pod-to-pod      | other-node | TCP_STREAM      | 10s             | 9541.04      
📋 host-to-host    | other-node | TCP_STREAM      | 10s             | 9579.99      
-------------------------------------------------------------------------------------
```
Previously native throughput was showing sad 1500 Mb/s for the same Cilium configuration. 

How the new output looks like in perfdash:
![image](https://github.com/user-attachments/assets/01d714c6-5e49-469f-9331-216d95e22c82)
![image](https://github.com/user-attachments/assets/e64a30bb-d379-4a92-8877-c4917bc306cd)
Makes it much easier to compare host vs pod throughput / latency etc.