@marseel 2024-08-12
"Lock cilium-cli repo / Lock cilium-cli repo (pull_request)"
:sob: